pyproject.toml
[build-system]
requires = ["wheel", "setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "naapc"
version = "2.1.0"
description = "Nested Automated Argument Parsing Configuration (NAAPC)."
readme = "README.md"
authors = [{name = "Bai Huanyu", email = "eiphnix@gmail.com"}]
license = {file = "LICENSE"}

classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["configuration", "config", "dictionary", "nested", "argument parsing"]
dependencies = ["pyyaml"]
requires-python = "==3.10.*"

[project.urls]
repository = "https://github.com/eiphy/naapc"

[tool.bumpver]
current_version = "2.1.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = false
tag = false
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"', 'version = "{version}"'
]
"src/naapc/__init__.py" = ["{version}"]
